+++ Reporting/SegmentSummary.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SmartSeg.Reporting
+{
+    public class SegmentSummary
+    {
+        public int ClusterId { get; set; }
+        public string Description { get; set; }
+        public Dictionary<string, string> Stats { get; set; }
+    }
+}
